cambios en vistas 2
