+++ src/front/js/layout.js
@@ -3,7 +3,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import ScrollToTop from "./component/scrollToTop.js";
 import { BackendURL } from "./component/backendURL.js";
 
-import { Home } from "./pages/home.js";
+import { Home } from "./pages/courses.js";
 import { Login } from "./pages/login.js";
 import { Payment } from "./pages/payment.js";
 import { Blog } from "./pages/blog.js";
@@ -26,7 +26,8 @@ const Layout = () => {
                 <ScrollToTop>
                     <Navbar />
                     <Routes>
-                        <Route element={<Home />} path="/" />
+                    <Route element={<Home />} path="/" />
+                        <Route element={<Courses />} path="/courses" />
                         <Route element={<Login />} path="/login" />
                         <Route element={<Blog />} path="/blog" />
                         <Route element={<Payment />} path="/payment/:theid" />
