+++ src/front/js/pages/payment.js
@@ -6,23 +6,8 @@ import rigoImageUrl from "../../img/rigo-baby.jpg";
 
 export const Payment = props => {
 	const { store, actions } = useContext(Context);
-	const params = useParams();
 
 	return (
-		<div className="jumbotron">
-			<h1 className="display-4">This will show the demo element: {store.demo[params.theid].title}</h1>
-			<img src={rigoImageUrl} />
-			<hr className="my-4" />
-
-			<Link to="/">
-				<span className="btn btn-primary btn-lg" href="#" role="button">
-					Back home
-				</span>
-			</Link>
-		</div>
+		<h1>Esto es la página del carrito</h1>
 	);
-};
-
-Single.propTypes = {
-	match: PropTypes.object
 };
